Update RequestHandler.ts
No longer need to import { URL } from "url" after node version 10 or above. The class is available on the global object after v10.0.0.
If import, it will cause error like: Uncaught (in promise) TypeError: xxxxxx.URL is not a constructor. 
I think this is an important update. I cannot use all of the functions without removing this import!